// client/generated-api/models/OwnNoteDto.ts
// tslint:disable
/**
 * Lappuzet
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface OwnNoteDto
 */
export interface OwnNoteDto {
    /**
     * @type {string}
     * @memberof OwnNoteDto
     */
    created_at: string;
    /**
     * @type {number}
     * @memberof OwnNoteDto
     */
    id: number;
    /**
     * @type {string}
     * @memberof OwnNoteDto
     */
    message: string;
    /**
     * @type {boolean}
     * @memberof OwnNoteDto
     */
    resolved: boolean;
    /**
     * @type {string}
     * @memberof OwnNoteDto
     */
    resolved_at: string;
}