// client/generated-api/models/NoteCategoryDto.ts
// tslint:disable
/**
 * Lappuzet
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface NoteCategoryDto
 */
export interface NoteCategoryDto {
    /**
     * @type {string}
     * @memberof NoteCategoryDto
     */
    created_at: string;
    /**
     * @type {number}
     * @memberof NoteCategoryDto
     */
    id: number;
    /**
     * @type {string}
     * @memberof NoteCategoryDto
     */
    name: string;
}